requirements.txt
asgiref==3.8.1
Django==3.2.12
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.3.0
pytz==2024.1
sqlparse==0.5.0
whitenoise==6.6.0
